asset/dcr: Shutdown RPC client in go routine.
Because the client does not accept a context or error on a disconnection,
the app will not shut down properly, waiting for the request to return
indefinitely. This will shut down the RPC client in a separate go
routine which will cause the client to return an error and the app to
continue shutting down.